+++ src/index.tsx
@@ -2,4 +2,5 @@ export * from './components/ContainerLink';
 export * from './components/ProgressLoader';
 export * from './types';
 
-export { changeRoute, eventEmitter } from './services/eventsServices';
+export { changeRoute } from './services/eventsServices';
+export { useRouter } from './hooks/useRouter';